HZN-1219: Read coordinates as double instead of float

+++ features/geolocation/api/src/main/java/org/opennms/features/geolocation/api/Coordinates.java
@@ -29,27 +29,27 @@
 package org.opennms.features.geolocation.api;
 
 public class Coordinates {
-    private float longitude;
-    private float latitude;
+    private double longitude;
+    private double latitude;
 
-    public Coordinates(float longitude, float latitude) {
+    public Coordinates(double longitude, double latitude) {
         this.longitude = longitude;
         this.latitude = latitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 }
