+++ features/geolocation/service/src/main/java/org/opennms/features/geolocation/services/DefaultGeolocationService.java
@@ -74,7 +74,7 @@ public class DefaultGeolocationService implements GeolocationService {
         final List<GeolocationInfo> nodesWithCoordinates = nodes.stream()
                 .filter(n -> geoLocation(n) != null && geoLocation(n).getLongitude() != null && geoLocation(n).getLatitude() != null)
                 // Avoid including -inf values, just in case. See NMS-9338
-                .filter(n -> geoLocation(n).getLatitude() != Float.NEGATIVE_INFINITY && geoLocation(n).getLongitude() != Float.NEGATIVE_INFINITY)
+                .filter(n -> geoLocation(n).getLatitude() != Double.NEGATIVE_INFINITY && geoLocation(n).getLongitude() != Double.NEGATIVE_INFINITY)
                 .map(node -> convert(node))
                 .collect(Collectors.toList());
